Massive refactoring

+++ lib/ascii-engine/display/canvas.cpp
@@ -1,6 +1,6 @@
 #include <sstream>
 #include <string>
-#include "ascii-engine/display/canvas.h"
+#include "canvas.h"
 
 namespace ae = ascii_engine;
 using namespace std;
@@ -11,7 +11,7 @@ void ae::Canvas::draw(int x, int y, const string& s) {
     while (iss.good() && !s.empty() && j < buffer.get_height()) {
         string line;
         getline(iss, line);
-        for (int i = x; i < buffer.get_width() && i - x < line.length(); ++i) {
+        for (unsigned i = x; i < (unsigned)buffer.get_width() && i - x < line.length(); ++i) {
             if (i < 0 || j < 0) {
                 continue;
             }
